+++ harfbuzz/src/hb-config.hh
@@ -33,64 +33,7 @@
 
 #include "config.h"
 
-
-#ifdef HB_TINY
-#define HB_LEAN
-#define HB_MINI
-#define HB_NO_MT
-#define HB_NO_UCD_UNASSIGNED
-#ifndef NDEBUG
-#define NDEBUG
-#endif
-#ifndef __OPTIMIZE_SIZE__
-#define __OPTIMIZE_SIZE__
-#endif
-#endif
-
-#ifdef HB_LEAN
-#define HB_DISABLE_DEPRECATED
-#define HB_NDEBUG
-#define HB_NO_ATEXIT
-#define HB_NO_BUFFER_MESSAGE
-#define HB_NO_BUFFER_SERIALIZE
-#define HB_NO_BITMAP
-#define HB_NO_CFF
-#define HB_NO_COLOR
-#define HB_NO_ERRNO
-#define HB_NO_FACE_COLLECT_UNICODES
-#define HB_NO_GETENV
-#define HB_NO_HINTING
-#define HB_NO_LANGUAGE_PRIVATE_SUBTAG
-#define HB_NO_LAYOUT_FEATURE_PARAMS
-#define HB_NO_LAYOUT_COLLECT_GLYPHS
-#define HB_NO_LAYOUT_UNUSED
-#define HB_NO_MATH
-#define HB_NO_META
-#define HB_NO_METRICS
-#define HB_NO_MMAP
-#define HB_NO_NAME
-#define HB_NO_OPEN
-#define HB_NO_SETLOCALE
-#define HB_NO_OT_FONT_GLYPH_NAMES
-#define HB_NO_OT_SHAPE_FRACTIONS
-#define HB_NO_STAT
-#define HB_NO_SUBSET_LAYOUT
-#define HB_NO_VAR
-#endif
-
-#ifdef HB_MINI
-#define HB_NO_AAT
-#define HB_NO_LEGACY
-#endif
-
-
 /* Closure of options. */
-
-#ifdef HB_DISABLE_DEPRECATED
-#define HB_IF_NOT_DEPRECATED(x)
-#else
-#define HB_IF_NOT_DEPRECATED(x) x
-#endif
 
 #ifdef HB_NO_AAT
 #define HB_NO_OT_NAME_LANGUAGE_AAT
@@ -99,11 +42,6 @@
 
 #ifdef HB_NO_BITMAP
 #define HB_NO_OT_FONT_BITMAP
-#endif
-
-#ifdef HB_NO_CFF
-#define HB_NO_OT_FONT_CFF
-#define HB_NO_SUBSET_CFF
 #endif
 
 #ifdef HB_NO_GETENV
@@ -118,21 +56,6 @@
 #define HB_NO_OT_SHAPE_FALLBACK
 #endif
 
-#ifdef HB_NO_NAME
-#define HB_NO_OT_NAME_LANGUAGE
-#endif
-
-#ifdef HB_NO_OT
-#define HB_NO_OT_FONT
-#define HB_NO_OT_LAYOUT
-#define HB_NO_OT_TAG
-#define HB_NO_OT_SHAPE
-#endif
-
-#ifdef HB_NO_OT_SHAPE
-#define HB_NO_AAT_SHAPE
-#endif
-
 #ifdef HB_NO_OT_SHAPE_FALLBACK
 #define HB_NO_OT_SHAPE_COMPLEX_ARABIC_FALLBACK
 #define HB_NO_OT_SHAPE_COMPLEX_HEBREW_FALLBACK
@@ -145,16 +68,5 @@
 #define HB_NDEBUG
 #endif
 #endif
-
-#ifdef __OPTIMIZE_SIZE__
-#ifndef HB_OPTIMIZE_SIZE
-#define HB_OPTIMIZE_SIZE
-#endif
-#endif
-
-#ifdef HAVE_CONFIG_OVERRIDE_H
-#include "config-override.h"
-#endif
-
 
 #endif /* HB_CONFIG_HH */